Get thumbnail programmatically instead of rendering from template.
