+++ topnotchdev/files_widget/templatetags/files_widget_tags.py
@@ -7,15 +7,6 @@ from django import template
 register = template.Library()
 
 @register.filter
-def thumbnail_format(path):
-    match = re.search(r'\.\w+$', path)
-    if match:
-        ext = match.group(0)
-        if ext.lower() in ['.gif', '.png']:
-            return 'PNG'
-    return 'JPEG'
-
-@register.filter
 def filename_from_path(path):
     return re.sub(r'^.+\/', '', path)
